Adds another soy file with a missing feature
Passing params with html content to other template calls doens't work right now. The compiled file correctly avoids running the html before passing it to the template, by wrapping it into a function. The template that gets the html param receives it as a function then, but it passes it directly to an `itext` call instead of calling it, so the function gets printed as text instead of actually rendering its markup.